guest_user

+++ app/controllers/users_controller.rb
@@ -7,12 +7,12 @@ class UsersController < ApplicationController
 
   def index
     @users = User.order(id: :desc).page(params[:page])
-  end 
-  
+  end
+
   def new
     @user = User.new
-  end 
-  
+  end
+
   def create
     @user = User.new(user_params)
     if @user.save
@@ -23,56 +23,66 @@ class UsersController < ApplicationController
       flash[:danger] = "ユーザーの登録に失敗しました"
       render(:new)
     end
-  end 
-  
+  end
+
   def show
     @tweets = @user.tweets.order(id: :desc).page(params[:page]).per(6)
     count(@user)
-  end 
-  
+  end
+
   def edit
-  end 
-  
+  end
+
   def update
-    if @user.update(user_params)
-      flash[:success] = "ユーザー情報を編集しました"
-      redirect_to(user_path)
+    if @user.email == "guest@guest.guest"
+      flash[:danger] = "ゲストユーザーのメールアドレス、パスワードは変更できません"
+      redirect_to(@user)
     else
-      flash[:danger] = "ユーザー情報の編集に失敗しました"
-      render(:edit)
+      if @user.update(user_params)
+        flash[:success] = "ユーザー情報を編集しました"
+        redirect_to(user_path)
+      else
+        flash[:danger] = "ユーザー情報の編集に失敗しました"
+        render(:edit)
+      end
     end
-  end 
-  
+  end
+
   def destroy
-    current_user.destroy
-    flash[:success] = "ユーザー情報を削除しました"
-    redirect_to(signup_path)
-  end 
-  
+    if current_user.email == "guest@guest.guest"
+      flash[:danger] = "ゲストユーザーは退会できません"
+      redirect_to(current_user)
+    else
+      current_user.destroy
+      flash[:success] = "ユーザー情報を削除しました"
+      redirect_to(signup_path)
+    end
+  end
+
   def destroy_image
     @user.remove_image!
     @user.save
     redirect_to(@user)
   end
-  
+
   def followings
     @followings = @user.followings.order(id: :desc).page(params[:page])
     count(@user)
-  end 
-  
+  end
+
   def followers
     @followers = @user.followers.order(id: :desc).page(params[:page])
     count(@user)
-  end 
-  
+  end
+
   def likes
     @likes = @user.favorite_tweets.order(id: :desc).page(params[:page]).per(6)
     count(@user)
   end
-  
+
   def edit_profile
-  end 
-  
+  end
+
   def update_profile
     if @user.update(profile_params)
       flash[:success] = "プロフィール情報を編集しました"
@@ -81,22 +91,22 @@ class UsersController < ApplicationController
       flash[:danger] = "プロフィール情報の編集に失敗しました"
       render(:edit_profile)
     end
-  end 
-  
+  end
+
  private
- 
+
  def user_params
    params.require(:user).permit(:name, :email, :password, :password_confirmation)
  end
- 
+
  def profile_params
    params.require(:user).permit(:image, :profile, :training_years, :age, :weight, :sex)
  end
- 
+
  def set_user
    @user = User.find(params[:id])
- end 
- 
+ end
+
  def correct_user
    if @user != @current_user
      flash[:danger] = "権限がありません"
